+++ src/model/column.rs
@@ -8,9 +8,9 @@ pub struct Column {
 }
 
 impl Column {
-    pub fn new(length: usize) -> Column {
+    pub fn new(length: i32) -> Column {
         let mut lines = Vec::new();
-        lines.resize_with(length, Default::default);
+        lines.resize_with(length as usize, Default::default);
         Column {
             lines,
         }
@@ -20,18 +20,18 @@ impl Column {
         self.lines.iter()
     }
 
-    pub fn len(&self) -> usize {
-        self.lines.len()
+    pub fn len(&self) -> i32 {
+        self.lines.len() as i32
     }
 
-    pub fn move_cursor(&self, local_x_cursor: usize, cursor_y: usize, direction: Direction) -> i32 {
+    pub fn move_cursor(&self, local_x_cursor: i32, cursor_y: i32, direction: Direction) -> i32 {
         match direction {
-            Direction::Left | Direction::Right => self.lines[cursor_y].move_cursor(local_x_cursor, direction),
+            Direction::Left | Direction::Right => self.lines[cursor_y as usize].move_cursor(local_x_cursor, direction),
             _ => panic!("This function should not be called with this direction")
         }
     }
 
-    pub fn line_mut(&mut self, index: usize) -> &mut ColumnLine {
-        &mut self.lines[index]
+    pub fn line_mut(&mut self, index: i32) -> &mut ColumnLine {
+        &mut self.lines[index as usize]
     }
 }
