+++ src/view/column.rs
@@ -10,8 +10,8 @@ use crate::view::Draw;
 #[derive(new)]
 pub struct ColumnDrawData {
     local_x_cursor: i32,
-    cursor_y: usize,
-    from_line: usize,
+    cursor_y: i32,
+    from_line: i32,
 }
 
 impl Draw for Column {
@@ -19,12 +19,12 @@ impl Draw for Column {
 
     fn draw(&self, renderer: &mut Renderer, x: i32, mut y: i32, theme: &Theme, ColumnDrawData { local_x_cursor, cursor_y, from_line }: ColumnDrawData) {
         let nb_line = self.len() - from_line;
-        let pattern_background_width = renderer.glyph_width() * (ColumnLineElement::LINE_LEN + ColumnLineElement::SIZE - 1) as u32;
-        let pattern_background_height = renderer.glyph_height() * nb_line as u32;
+        let pattern_background_width = renderer.glyph_width() * (ColumnLineElement::LINE_LEN + ColumnLineElement::SIZE as i32 - 1);
+        let pattern_background_height = renderer.glyph_height() * nb_line;
         renderer.draw_rect(x, y, pattern_background_width, pattern_background_height, theme.pattern_background_color());
-        for (line_index, line) in self.iter().enumerate().skip(from_line) {
-            line.draw(renderer, x, y, theme, ColumnLineDrawData::new(line_index == cursor_y, local_x_cursor));
-            y += renderer.glyph_height() as i32;
+        for (line_index, line) in self.iter().enumerate().skip(from_line as usize) {
+            line.draw(renderer, x, y, theme, ColumnLineDrawData::new(line_index as i32 == cursor_y, local_x_cursor));
+            y += renderer.glyph_height();
         }
     }
 }
