+++ src/view/column_line.rs
@@ -19,13 +19,13 @@ impl Draw for ColumnLine {
 
     fn draw(&self, renderer: &mut Renderer, mut x: i32, y: i32, theme: &Theme, ColumnLineDrawData { is_active_line, local_x_cursor }: ColumnLineDrawData) {
         if is_active_line {
-            let gray_highlight_width = renderer.glyph_width() * (ColumnLineElement::LINE_LEN + ColumnLineElement::SIZE - 1) as u32;
+            let gray_highlight_width = renderer.glyph_width() * (ColumnLineElement::LINE_LEN + ColumnLineElement::SIZE as i32 - 1);
             renderer.draw_rect(x, y, gray_highlight_width, renderer.glyph_height(), theme.highlighted_background_color());
         }
 
         self.note.draw(renderer, x, y, theme, NoteFieldDrawData::new(if is_active_line { Some(local_x_cursor) } else { None }));
 
-        x += renderer.glyph_width() as i32 * (ColumnLineElement::Note.len() + 1) as i32;
+        x += renderer.glyph_width() * (ColumnLineElement::Note.len() + 1) as i32;
 
         self.velocity.draw(renderer, x, y, theme, VelocityFieldDrawData::new(if is_active_line { Some(local_x_cursor - 3) } else { None }));
     }
