+++ src/view/patterns.rs
@@ -8,7 +8,7 @@ use crate::view::pattern::PatternDrawData;
 impl Draw for Patterns {
     fn draw(&self, renderer: &mut Renderer, x: i32, mut y: i32, theme: &Theme, _: ()) {
         renderer.draw_text_with_background(format!("Pattern {}/{}", self.selected_pattern_index + 1, self.nb_patterns()), x, y, theme.text_color(), theme.pattern_background_color(), TextAlignment::Left);
-        y += renderer.glyph_height() as i32;
+        y += renderer.glyph_height();
         self.current_pattern().draw(renderer, x, y, theme, PatternDrawData::new(self.cursor_x, self.cursor_y));
     }
 }
